# server/docs/swagger.yaml
basePath: /
definitions:
  auth.LoginRequest:
    properties:
      captchaId:
        type: string
      captchaValue:
        type: string
      password:
        example: admin123
        type: string
      userName:
        example: admin
        type: string
      username:
        example: admin
        type: string
    type: object
...
  response.Response:
    properties:
      code: {}
      data: {}
      msg:
        type: string
      request_id:
        type: string
    type: object
...
  system.ChangePasswordReq:
    properties:
      newPassword:
        maxLength: 32
        minLength: 8
        type: string
      oldPassword:
        type: string
    required:
    - newPassword
    - oldPassword
    type: object
...
  system.CreateAdminReq:
    properties:
      email:
        maxLength: 100
        type: string
      gender:
        enum:
        - "0"
        - "1"
        - "2"
        type: string
      nickname:
        type: string
      password:
        maxLength: 32
        minLength: 8
        type: string
      phone:
        type: string
      roles:
        items:
          type: string
        type: array
      status:
        enum:
        - "0"
        - "1"
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - password
    - status
    - username
    type: object
...
  system.RefreshTokenReq:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
...
  system.UpdateAdminReq:
    properties:
      email:
        maxLength: 100
        type: string
      gender:
        enum:
        - "0"
        - "1"
        - "2"
        type: string
      id:
        type: integer
      nickname:
        type: string
      password:
        maxLength: 32
        minLength: 8
        type: string
      phone:
        type: string
      roles:
        items:
          type: string
        type: array
      status:
        enum:
        - "0"
        - "1"
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - status
    - username
    type: object
...
  system.UpdateProfileReq:
    properties:
      email:
        maxLength: 100
        type: string
      gender:
        enum:
        - "0"
        - "1"
        - "2"
        type: string
      nickname:
        type: string
      phone:
        type: string
    type: object
...
info:
  contact: {}
  description: NetyAdmin 后台管理系统 API 文档
  title: NetyAdmin API
  version: "1.0"
paths:
  /admin/v1/admins:
    get:
      consumes:
      - application/json
      description: 分页查询管理员列表，支持按用户名、昵称、手机号、邮箱、状态、性别筛选
      parameters:
      - description: 用户名
        in: query
        name: username
        type: string
      - description: 昵称
        in: query
        name: nickname
        type: string
      - description: 手机号
        in: query
        name: phone
        type: string
      - description: 邮箱
        in: query
        name: email
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - description: 性别
        in: query
        name: gender
        type: string
      - description: 当前页
        in: query
        name: current
        type: integer
      - description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 管理员分页列表
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取管理员列表
      tags:
      - 管理员管理
    post:
      consumes:
      - application/json
      description: 创建新的管理员账号，设置用户名、密码、昵称、联系方式、状态及角色
      parameters:
      - description: 管理员创建参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/system.CreateAdminReq'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建管理员
      tags:
      - 管理员管理
  /admin/v1/admins/{id}:
    delete:
      consumes:
      - application/json
      description: 根据管理员ID删除单个管理员账号
      parameters:
      - description: 管理员ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除管理员
      tags:
      - 管理员管理
    get:
      consumes:
      - application/json
      description: 根据管理员ID获取管理员详细信息
      parameters:
      - description: 管理员ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 管理员详情
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取管理员详情
      tags:
      - 管理员管理
    put:
      consumes:
      - application/json
      description: 根据管理员ID更新管理员账号信息，包括资料、密码及角色
      parameters:
      - description: 管理员ID
        in: path
        name: id
        required: true
        type: integer
      - description: 管理员更新参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/system.UpdateAdminReq'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新管理员
      tags:
      - 管理员管理
  /admin/v1/admins/batch:
    delete:
      consumes:
      - application/json
      description: 根据管理员ID列表批量删除管理员账号
      parameters:
      - description: 待删除的管理员ID列表
        in: body
        name: ids
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 批量删除成功
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 批量删除管理员
      tags:
      - 管理员管理
  /admin/v1/auth/change-password:
    post:
      consumes:
      - application/json
      description: 当前登录管理员通过旧密码修改登录密码
      parameters:
      - description: 修改密码请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/system.ChangePasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: 修改成功
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 修改密码
      tags:
      - 认证管理
  /admin/v1/auth/user-info:
    get:
      consumes:
      - application/json
      description: 根据上下文中的管理员ID获取当前登录用户的详细信息
      produces:
      - application/json
      responses:
        "200":
          description: 用户信息
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取当前登录用户信息
      tags:
      - 认证管理
  /admin/v1/auth/login:
    post:
      consumes:
      - application/json
      description: 使用用户名/密码登录管理员账号，可选验证码校验，返回访问令牌与刷新令牌
      parameters:
      - description: 登录参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            $ref: '#/definitions/response.Response'
      summary: 管理员登录
      tags:
      - 认证管理
  /admin/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: 当前登录管理员退出登录，注销访问令牌
      produces:
      - application/json
      responses:
        "200":
          description: 退出成功
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 退出登录
      tags:
      - 认证管理
  /admin/v1/auth/profile:
    get:
      consumes:
      - application/json
      description: 获取当前登录管理员的个人资料信息
      produces:
      - application/json
      responses:
        "200":
          description: 个人资料
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取个人资料
      tags:
      - 认证管理
    put:
      consumes:
      - application/json
      description: 更新当前登录管理员的个人资料，例如昵称、手机号、邮箱、性别
      parameters:
      - description: 个人资料信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/system.UpdateProfileReq'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新个人资料
      tags:
      - 认证管理
  /admin/v1/auth/refresh-token:
    post:
      consumes:
      - application/json
      description: 使用刷新令牌换取新的访问令牌与刷新令牌
      parameters:
      - description: 刷新令牌请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/system.RefreshTokenReq'
      produces:
      - application/json
      responses:
        "200":
          description: 刷新成功
          schema:
            $ref: '#/definitions/response.Response'
      summary: 刷新令牌
      tags:
      - 认证管理
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"