# server/internal/interface/admin/dto/system/admin.go
package system

import "NetyAdmin/internal/interface/admin/dto"

type LoginReq struct {
	Username string `json:"username" binding:"required"`
	Password string `json:"password" binding:"required"`
}

type RefreshTokenReq struct {
	RefreshToken string `json:"refreshToken" binding:"required"`
}

type ChangePasswordReq struct {
	OldPassword string `json:"oldPassword" binding:"required"`
	NewPassword string `json:"newPassword" binding:"required,min=8,max=32"`
}

type UpdateProfileReq struct {
	Nickname string `json:"nickname"`
	Phone    string `json:"phone"`
	Email    string `json:"email" binding:"omitempty,email,max=100"`
	Gender   string `json:"gender" binding:"omitempty,oneof=0 1 2"`
}

type AdminQuery struct {
	dto.PageQuery
	Username string  `form:"username" json:"username"`
	Nickname string  `form:"nickname" json:"nickname"`
	Phone    string  `form:"phone" json:"phone"`
	Email    string  `form:"email" json:"email" binding:"omitempty,email,max=100"`
	Status   *string `form:"status" json:"status"`
	Gender   *string `form:"gender" json:"gender" binding:"omitempty,oneof=0 1 2"`
}

type CreateAdminReq struct {
	Username string   `json:"username" binding:"required,min=3,max=50"`
	Password string   `json:"password" binding:"required,min=8,max=32"`
	Nickname string   `json:"nickname"`
	Phone    string   `json:"phone"`
	Email    string   `json:"email" binding:"omitempty,email,max=100"`
	Gender   string   `json:"gender" binding:"omitempty,oneof=0 1 2"`
	Status   string   `json:"status" binding:"required,oneof=0 1"`
	Roles    []string `json:"roles"`
}

type UpdateAdminReq struct {
	ID       uint     `json:"id"`
	Username string   `json:"username" binding:"required,min=3,max=50"`
	Password string   `json:"password" binding:"omitempty,min=8,max=32"`
	Nickname string   `json:"nickname"`
	Phone    string   `json:"phone"`
	Email    string   `json:"email" binding:"omitempty,email,max=100"`
	Gender   string   `json:"gender" binding:"omitempty,oneof=0 1 2"`
	Status   string   `json:"status" binding:"required,oneof=0 1"`
	Roles    []string `json:"roles"`
}

// BatchDeleteAdminReq 批量删除管理员请求
type BatchDeleteAdminReq struct {
	IDs []uint `json:"ids" binding:"required,min=1"`
}